Adjusted bit_vector from 'long int' to 'long long'

+++ tbamud-3.64/src/structs.h
@@ -914,7 +914,7 @@ typedef IDXTYPE qst_rnum;   /**< references an instance of a quest */
  * will give you at least 64 bits if you have GCC. You'll have to search
  * throughout the code for "bitvector_t" and change them yourself if you'd
  * like this extra flexibility. */
-typedef unsigned long int bitvector_t;
+typedef unsigned long long bitvector_t;
 
 /** Extra description: used in objects, mobiles, and rooms. For example,
  * a 'look hair' might pull up an extra description (if available) for
